# docs/swagger.yaml
consumes:
- application/json
definitions:
...
  http.AdminAssignRoleRequest:
    properties:
      role:
        example: moderator
        type: string
    required:
    - role
    type: object
  http.AdminDiseaseCategoryUpsertRequest:
    properties:
      title:
        example: Cashflow
        type: string
    required:
    - title
    type: object
...
  http.AdminUserRolesResponse:
    properties:
      roles:
        example:
        - participant
        - moderator
        items:
          type: string
        type: array
      userId:
        example: cbedcee0-1673-432d-b5b9-2a35aabae68c
        type: string
    type: object
...
  http.DiseaseCategoryResponse:
    properties:
      code:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: API for Business Surgery service.
  title: Business Surgery Backend API
  version: "1.0"
paths:
...
  /api/v1/admin/disease-categories:
    get:
      parameters:
      - description: Bearer {accessToken}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.DiseaseCategoryResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List disease categories (admin)
      tags:
      - admin-categories
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer {accessToken}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Category payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.AdminDiseaseCategoryUpsertRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.DiseaseCategoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create disease category
      tags:
      - admin-categories
  /api/v1/admin/disease-categories/{id}:
    delete:
      description: Deletes category if it is not referenced by any disease.
      parameters:
      - description: Bearer {accessToken}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Category ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete disease category
      tags:
      - admin-categories
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer {accessToken}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Category ID (uuid)
        in: path
        name: id
        required: true
        type: string
      - description: Category payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.AdminDiseaseCategoryUpsertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.DiseaseCategoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update disease category
      tags:
      - admin-categories
...
  /api/v1/admin/users/{id}/roles:
    get:
      description: Returns all roles assigned to a user
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.AdminUserRolesResponse'
        "400":
          description: invalid user id
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: user not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user roles
      tags:
      - admin-rbac
    post:
      consumes:
      - application/json
      description: Adds a role to the user (does not remove existing roles)
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Role to assign
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.AdminAssignRoleRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: user or role not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: role already assigned
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Assign role to user
      tags:
      - admin-rbac
  /api/v1/admin/users/{id}/roles/{roleCode}:
    delete:
      description: Removes a role from the user. Cannot remove the last remaining
        role.
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Role code (participant | moderator | admin)
        in: path
        name: roleCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: invalid input or last role
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: role not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove role from user
      tags:
      - admin-rbac
...
produces:
- application/json
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"